requirements.txt
django==5.1.5
djangorestframework==3.15.2
django-environ==0.12.0
djangorestframework==simplejwt-5.4.0
mysqlclient==2.2.7